Fix eyedropper/move tool for reference layers

+++ src/doc/sprite.h
@@ -140,7 +140,7 @@ namespace doc {
     void replaceImage(ObjectId curImageId, const ImageRef& newImage);
     void getImages(std::vector<Image*>& images) const;
     void remapImages(frame_t frameFrom, frame_t frameTo, const Remap& remap);
-    void pickCels(int x, int y, frame_t frame, int opacityThreshold, CelList& cels) const;
+    void pickCels(double x, double y, frame_t frame, int opacityThreshold, CelList& cels) const;
 
     ////////////////////////////////////////
     // Iterators
